+++ tests/sdk/services/test_model_factory.py
@@ -1,8 +1,8 @@
 from unittest.mock import Mock, patch
 
 import pytest
-from rhesis.sdk.services.base import BaseLLM
-from rhesis.sdk.services.model_factory import (
+from rhesis.sdk.models.base import BaseLLM
+from rhesis.sdk.models.model_factory import (
     DEFAULT_MODELS,
     DEFAULT_PROVIDER,
     ModelConfig,
